+++ src/CameraAndor.h
@@ -73,7 +73,7 @@ protected:
 	/*!
 	 * @brief 改变制冷状态和制冷温度
 	 */
-	bool update_cooler(double coolerset, bool onoff);
+	bool update_cooler(int coolerset, bool onoff);
 	/*!
 	 * @brief 采集探测器温度
 	 * @param coolerget 探测器温度
